\name{marijuana}
\docType{data}
\alias{marijuana}
\title{A pilot study of the clinical and psychological effects of maarijuana}
\description{
Nine male subjects were given three treatments in the form of low-dose, high-dose,
and placebo cigarettes. The order of treatments within subjects was balanced in a
replicated 3 x 3 Latin square, but because the order for each subject was not
reported in the article, (I shall proceed as if) the order effects are negligible.
Changes in heart rate were recorded 15 and 90 minutes after marijuana use, and
five of the 54 data values are missing.
}
\usage{data(marijuana)}
\format{A data frame}
\source{Wei AT, Zinberg NE, Nelson JM. Clinical and psychological effects of 
marijuana in man. Science 1968; 162:1234-1242}
  
\keyword{datasets}
